[Silver V] Title: 수들의 합, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1789. 수들의 합/수들의 합.cc
@@ -5,7 +5,7 @@ void ans(){
     long S;
     cin >> S;
 
-    long N = 0;
+    int N = 0;
 
     while(S > 0){
         N++;
